+++ buraq/contrib/email/__init__.py
@@ -1,5 +1,6 @@
 from buraq.contrib.email.message import EmailMessage, EmailMultiAlternatives
 from buraq.contrib.email.send import (
+    get_connection,
     mail_admins,
     mail_managers,
     send_mail,
@@ -10,6 +11,7 @@ from buraq.contrib.email.send import (
 __all__ = [
     "EmailMessage",
     "EmailMultiAlternatives",
+    "get_connection",
     "send_mail",
     "send_mass_mail",
     "send_template_mail",